HW3/requirements.txt
aiohttp==3.11.14
aiosqlite==0.21.0
alembic==1.15.1
asyncpg==0.30.0
celery==5.4.0
fastapi==0.115.11
fastapi_cache==0.1.0
fastapi-cache2==0.2.2
fastapi-users[sqlalchemy]==14.0.1
flower==2.0.1
httpx==0.28.1
locust==2.33.2
psycopg2-binary
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.0.1
redis==5.2.1
sqlalchemy==2.0.39
trio==0.29.0
uvicorn==0.34.0
gunicorn